requirements.txt
aiogram~=3.3.0
python-decouple
requests~=2.31.0